ARROW-2725: [Java] make Accountant.AllocationOutcome publicly visible (#2149)
